moved Vehicle.java
