+++ services/shared/src/main/java/com/wedriveu/services/shared/entity/Vehicle.java
@@ -1,4 +1,4 @@
-package com.wedriveu.services.vehicle.entity;
+package com.wedriveu.services.shared.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.wedriveu.services.shared.utilities.Position;
